requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.1
propcache==0.4.1
pyTelegramBotAPI==4.26.0
python-dotenv==1.2.2
python-telegram-bot==21.10
requests==2.33.1
typing_extensions==4.15.0
urllib3==2.6.3
vt-py==0.22.0
yarl==1.23.0
